server: stream: new stream_input interfaces, arch changes (details below)
In the previous Stream interface, we used to connect multiple streams only
and attach them to a Channel:

      [         Channel         ]
                  |
      [stream IOV]->[stream file]

This patch changes the logic where the stream input concept is introduced,
summary:

 - A channel consumes a list of streams
 - Streams now are a group of inputs like: raw, iov, file, buf, etc.
 - Inputs are linked to a Stream only.

This new approach gives more control over data where is required to group
several sources that together represent an entity like a 'full response'.

This patch also introduce some basic API to implement web services, more work
is still required.

This patch breaks most of the plugins, just use it with Liana.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>