Plugins: cgi: fix effective user ID, drop privileges at fork(2)
This patch fix a security issue when using the CGI interface. When
Monkey is launched as root and the configuration key 'User' is set, it
switch to the given system user, so the Effective UID is changed.

The CGI plugin uses fork(2), when the child process is created it inherit
the UID as EUID, so we must drop privileges before invoking the target
script. This patch make sure to make the EUID/EGID becomes UID/GID.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>