server: experimental co-routine (threading) model for handlers (upon demand)
This patch implements the coroutines management so plugin handlers (stage30)
can request a context about it.

As of now it can create/resume/yield contexts. New network API calls under
mk_net.c works on top of coroutines so connect() works async. This is still
work in process and more interfaces are required to be implemented.

note: using fastcgi plugin for testing.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>