core: event: introduce mk_event_closesocket() function
Previously the downstream programs had to figure out the correct
function to close sockets depending on the backend library in use.
(i.e. use evutil_socketclose if the mk_core's backend is libevent)

Failing to do so would result in segmentation faults and it was
somewhat cumbersome. With this patch library users can now call
mk_event_closesocket() to close sockets safely.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>