event: libevent: suppress error messages on a normal shutdown
We notify the termination of an event loop to the timer worker by
closing the other end of the socket. This is detected as ECONNABORTED
by the worker after a failed call of write(2).

Since this is quite a normal operation, we do not need (and want) to
log it as an error.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>