core: winpthreads: fix segfaults on pthread_set_specific() (#302)
When pthread_key_create() generates the first TSD key, it fails
to initialize the first slot of _pthread_key_dest properly. For
this reason, the subsequent call for the method can generate an
invalid TSD key.

This was the root cause of fluent-bit randomly failing with
segmentation fault. Fix it.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>