rconf: add meaningful error message
When the content of one line in config file has exceed the value of the
Macro MK_RCONF_KV_SIZE, which is 4096 now. It will show error message
like "Invalid indentation level".  This PR has added an ELSE branch in
the code logical to give out a meaningful error message.

Signed-off-by: Fenggang <ginobiliwang@gmail.com>
Signed-off-by: Eduardo Silva <eduardo@monkey.io>