mk_core: mk_rconf: Fix potential overrun in mk_rconf_read_glob() (#308)
Justin Angra pointed out that these strcat() have a potential overrun
bug. This observation is correct. In particular, when the input string
`path` is very long, it can overrun the buffer while appending `p1`.

This patch fixes the issue by using a safer variant of strcat and
check the return code explicitly.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>