server: add support for transfer chunked encoding
When requests arrives through a POST or PUT method and uses
'chunked' transfer-encoding, the parser was not allowing or
processing the payload.

This patch extends the parser to support chunked transfer encoding
allowing the parser not to fail and validate the chunks that arrives.

It also provides some new helpers functions to decode and extract the
content in a new memory buffer. New functions available:

1. size_t mk_http_parser_content_length(struct mk_http_parser *p);

   indicate the content length of the payload, if content-length has been
   set in the headers it will return that length (after being validated). If
   the request came with chunked transfer encoding, it will return the total
   size of the chunks without including the chunked headers and \r\n marks.

2. int mk_http_parser_is_content_chunked(struct mk_http_parser *p);

   specify if the content being parsers contains chunked encoded data (bool).

3. int mk_http_parser_chunked_decode(struct mk_http_parser *p,
                                     char *buf_request, size_t buf_request_len,
                                     char **out_buf, size_t *out_buf_size);

   If the parser found chunked encoded data, and once the request has finished
   to be parsed, it returns a new memory buffer with the decoded data. The new
   buffer must be freed later with mk_mem_free() by the user.

Note that these functions can only be used once the first parsing state with
the function mk_http_parser() has returned MK_HTTP_PARSER_OK (which means that
the request is complete.

Signed-off-by: Eduardo Silva <eduardo@calyptia.com>